requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
psutil==5.8.0
pyobjc-core==7.1
pyobjc-framework-Cocoa==7.1
pyobjc-framework-Quartz==7.1
pyvesync==1.2.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
